MCP: Fix unit tests and move around files
