+++ src/experimental/tools/mediaCapabilitiesProber/probers/__tests__/HDCPPolicy.test.ts
@@ -12,9 +12,41 @@ describe("MediaCapabilitiesProber probers - HDCPPolicy", () => {
     }));
     const probeHDCPPolicy = (await vi.importActual("../../probers/HDCPPolicy"))
       .default as typeof IProbeHDCPPolicy;
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    expect(probeHDCPPolicy("")).rejects.toEqual(
-      "MediaCapabilitiesProber >>> API_CALL: API not available",
+    await probeHDCPPolicy("1.1").then(
+      () => {
+        throw new Error("Should not have succeeded");
+      },
+      (err: unknown) => {
+        expect(err).toBeInstanceOf(Error);
+      },
+    );
+  });
+
+  it("should reject if MediaKeys creation fails", async () => {
+    const mockCreateMediaKeys = vi.fn(() => {
+      return Promise.reject(new Error("NOPE LOL"));
+    });
+    const mockRequestMediaKeySystemAcces = vi.fn(() => {
+      return Promise.resolve({
+        createMediaKeys: mockCreateMediaKeys,
+      });
+    });
+    vi.doMock("../../../../../compat/eme", () => ({
+      default: {
+        requestMediaKeySystemAccess: mockRequestMediaKeySystemAcces,
+      },
+    }));
+
+    const probeHDCPPolicy = (await vi.importActual("../../probers/HDCPPolicy"))
+      .default as typeof IProbeHDCPPolicy;
+
+    await probeHDCPPolicy("1.1").then(
+      () => {
+        throw new Error("Should not have succeeded");
+      },
+      (err: unknown) => {
+        expect(err).toBeInstanceOf(Error);
+      },
     );
   });
 
@@ -36,7 +68,7 @@ describe("MediaCapabilitiesProber probers - HDCPPolicy", () => {
     const probeHDCPPolicy = (await vi.importActual("../../probers/HDCPPolicy"))
       .default as typeof IProbeHDCPPolicy;
 
-    await probeHDCPPolicy("1.1").then(([res]) => {
+    await probeHDCPPolicy("1.1").then((res: string) => {
       expect(res).toEqual("Unknown");
       expect(mockCreateMediaKeys).toHaveBeenCalledTimes(1);
       expect(mockRequestMediaKeySystemAcces).toHaveBeenCalledTimes(1);
@@ -63,7 +95,7 @@ describe("MediaCapabilitiesProber probers - HDCPPolicy", () => {
     const probeHDCPPolicy = (await vi.importActual("../../probers/HDCPPolicy"))
       .default as typeof IProbeHDCPPolicy;
 
-    await probeHDCPPolicy("1.1").then(([res]) => {
+    await probeHDCPPolicy("1.1").then((res: string) => {
       expect(res).toEqual("Supported");
       expect(mockCreateMediaKeys).toHaveBeenCalledTimes(1);
       expect(mockRequestMediaKeySystemAcces).toHaveBeenCalledTimes(1);
@@ -89,7 +121,7 @@ describe("MediaCapabilitiesProber probers - HDCPPolicy", () => {
 
     const probeHDCPPolicy = (await vi.importActual("../../probers/HDCPPolicy"))
       .default as typeof IProbeHDCPPolicy;
-    await probeHDCPPolicy("1.1").then(([res]) => {
+    await probeHDCPPolicy("1.1").then((res: string) => {
       expect(res).toEqual("NotSupported");
       expect(mockCreateMediaKeys).toHaveBeenCalledTimes(1);
       expect(mockRequestMediaKeySystemAcces).toHaveBeenCalledTimes(1);
