+++ src/experimental/tools/mediaCapabilitiesProber/probers/__tests__/decodingInfo.test.ts
@@ -51,28 +51,26 @@ describe("MediaCapabilitiesProber probers - decodingInfo", () => {
     navigator.mediaCapabilities = origMediaCapabilities;
   });
 
-  it("should throw if no video and audio config", () => {
+  it("should throw if no video and audio config", async () => {
     const decodingInfoStub = stubDecodingInfo(true);
     const configuration = {
       type: "media-source" as const,
     };
 
-    expect.assertions(2);
-    return probeDecodingInfos(configuration)
-      .then(() => {
-        resetDecodingInfos();
-      })
-      .catch(({ message }: { message: string }) => {
-        expect(message).toEqual(
-          "MediaCapabilitiesProber >>> API_CALL: " +
-            "Not enough arguments for calling mediaCapabilites.",
-        );
-        expect(decodingInfoStub).not.toHaveBeenCalled();
-        resetDecodingInfos();
-      });
+    let thrownException = false;
+    try {
+      await probeDecodingInfos(configuration);
+      resetDecodingInfos();
+    } catch ({ message }) {
+      thrownException = true;
+      expect(message).toEqual("Not enough arguments for calling mediaCapabilites.");
+      expect(decodingInfoStub).not.toHaveBeenCalled();
+      resetDecodingInfos();
+    }
+    expect(thrownException).toEqual(true);
   });
 
-  it("should throw if incomplete video config", () => {
+  it("should throw if incomplete video config", async () => {
     const decodingInfoStub = stubDecodingInfo(true);
     const configuration = {
       type: "media-source" as const,
@@ -81,22 +79,20 @@ describe("MediaCapabilitiesProber probers - decodingInfo", () => {
       },
     };
 
-    expect.assertions(2);
-    return probeDecodingInfos(configuration)
-      .then(() => {
-        resetDecodingInfos();
-      })
-      .catch(({ message }: { message: string }) => {
-        expect(message).toEqual(
-          "MediaCapabilitiesProber >>> API_CALL: " +
-            "Not enough arguments for calling mediaCapabilites.",
-        );
-        expect(decodingInfoStub).not.toHaveBeenCalled();
-        resetDecodingInfos();
-      });
+    let thrownException = false;
+    try {
+      await probeDecodingInfos(configuration);
+      resetDecodingInfos();
+    } catch ({ message }) {
+      thrownException = true;
+      expect(message).toEqual("Not enough arguments for calling mediaCapabilites.");
+      expect(decodingInfoStub).not.toHaveBeenCalled();
+      resetDecodingInfos();
+    }
+    expect(thrownException).toEqual(true);
   });
 
-  it("should throw if incomplete audio config", () => {
+  it("should throw if incomplete audio config", async () => {
     const decodingInfoStub = stubDecodingInfo(true);
     const configuration = {
       type: "media-source" as const,
@@ -105,68 +101,70 @@ describe("MediaCapabilitiesProber probers - decodingInfo", () => {
       },
     };
 
-    expect.assertions(2);
-    return probeDecodingInfos(configuration)
-      .then(() => {
-        resetDecodingInfos();
-      })
-      .catch(({ message }: { message: string }) => {
-        expect(message).toEqual(
-          "MediaCapabilitiesProber >>> API_CALL: " +
-            "Not enough arguments for calling mediaCapabilites.",
-        );
-        expect(decodingInfoStub).not.toHaveBeenCalled();
-        resetDecodingInfos();
-      });
+    let thrownException = false;
+    try {
+      await probeDecodingInfos(configuration);
+      resetDecodingInfos();
+    } catch ({ message }) {
+      thrownException = true;
+      expect(message).toEqual("Not enough arguments for calling mediaCapabilites.");
+      expect(decodingInfoStub).not.toHaveBeenCalled();
+      resetDecodingInfos();
+    }
+    expect(thrownException).toEqual(true);
   });
 
-  it("should throw if no type in config", () => {
+  it("should throw if no type in config", async () => {
     const decodingInfoStub = stubDecodingInfo(true);
     const configuration = {
       audio: {
         contentType: "audio/wma",
       },
     };
-    expect.assertions(2);
-    return probeDecodingInfos(configuration)
-      .then(() => {
-        resetDecodingInfos();
-      })
-      .catch(({ message }: { message: string }) => {
-        expect(message).toEqual(
-          "MediaCapabilitiesProber >>> API_CALL: " +
-            "Not enough arguments for calling mediaCapabilites.",
-        );
-        expect(decodingInfoStub).not.toHaveBeenCalled();
-        resetDecodingInfos();
-      });
+    let thrownException = false;
+    try {
+      await probeDecodingInfos(configuration);
+      resetDecodingInfos();
+    } catch ({ message }) {
+      thrownException = true;
+      expect(message).toEqual("Not enough arguments for calling mediaCapabilites.");
+      expect(decodingInfoStub).not.toHaveBeenCalled();
+      resetDecodingInfos();
+    }
+    expect(thrownException).toEqual(true);
   });
 
-  it("should throw if empty config", () => {
+  it("should throw if empty config", async () => {
     const decodingInfoStub = stubDecodingInfo(true);
     const configuration = {};
 
-    expect.assertions(2);
-    return probeDecodingInfos(configuration)
-      .then(() => {
-        resetDecodingInfos();
-      })
-      .catch(({ message }: { message: string }) => {
-        expect(message).toEqual(
-          "MediaCapabilitiesProber >>> API_CALL: " +
-            "Not enough arguments for calling mediaCapabilites.",
-        );
-        expect(decodingInfoStub).not.toHaveBeenCalled();
-        resetDecodingInfos();
-      });
+    let thrownException = false;
+    try {
+      await probeDecodingInfos(configuration);
+      resetDecodingInfos();
+    } catch ({ message }) {
+      thrownException = true;
+      expect(message).toEqual("Not enough arguments for calling mediaCapabilites.");
+      expect(decodingInfoStub).not.toHaveBeenCalled();
+      resetDecodingInfos();
+    }
+    expect(thrownException).toEqual(true);
   });
 
   it("should throw if API mediaCapabilities not available", async () => {
     // @ts-expect-error: `navigator.mediaCapabilities` is read-only normally, for
     // now, we're going through JSDom through so that's OK.
     delete navigator.mediaCapabilities;
-    await expect(probeDecodingInfos({})).rejects.toThrowError(
-      "MediaCapabilitiesProber >>> API_CALL: MediaCapabilities API not available",
+    return probeDecodingInfos({}).then(
+      () => {
+        throw new Error("Should not succeed");
+      },
+      (err: Error) => {
+        expect(err).toBeInstanceOf(Error);
+        expect(err.message).toEqual(
+          "navigator.mediaCapabilites.decodingInfo is not available",
+        );
+      },
     );
   });
 
@@ -180,12 +178,20 @@ describe("MediaCapabilitiesProber probers - decodingInfo", () => {
       // now, we're going through JSDom through so that's OK.
       navigator.mediaCapabilities = {};
     }
-    await expect(probeDecodingInfos({})).rejects.toThrowError(
-      "MediaCapabilitiesProber >>> API_CALL: Decoding Info not available",
+    return probeDecodingInfos({}).then(
+      () => {
+        throw new Error("Should not succeed");
+      },
+      (err: Error) => {
+        expect(err).toBeInstanceOf(Error);
+        expect(err.message).toEqual(
+          "navigator.mediaCapabilites.decodingInfo is not available",
+        );
+      },
     );
   });
 
-  it("should resolve with `Supported` if decodingInfo supports (video only)", () => {
+  it("should resolve with `Supported` if decodingInfo supports (video only)", async () => {
     const decodingInfoStub = stubDecodingInfo(true);
     const configuration: IMediaConfiguration = {
       type: "media-source",
@@ -198,19 +204,13 @@ describe("MediaCapabilitiesProber probers - decodingInfo", () => {
       },
     };
 
-    expect.assertions(2);
-    return probeDecodingInfos(configuration)
-      .then((res) => {
-        expect(res).toBe("Supported");
-        expect(decodingInfoStub).toHaveBeenCalledTimes(1);
-        resetDecodingInfos();
-      })
-      .catch(() => {
-        resetDecodingInfos();
-      });
+    const res = await probeDecodingInfos(configuration);
+    expect(res).toBe("Supported");
+    expect(decodingInfoStub).toHaveBeenCalledTimes(1);
+    resetDecodingInfos();
   });
 
-  it("should resolve with `Supported` if decodingInfo supports (audio only)", () => {
+  it("should resolve with `Supported` if decodingInfo supports (audio only)", async () => {
     const decodingInfoStub = stubDecodingInfo(true);
     const configuration: IMediaConfiguration = {
       type: "media-source",
@@ -222,19 +222,13 @@ describe("MediaCapabilitiesProber probers - decodingInfo", () => {
       },
     };
 
-    expect.assertions(2);
-    return probeDecodingInfos(configuration)
-      .then((res) => {
-        expect(res).toBe("Supported");
-        expect(decodingInfoStub).toHaveBeenCalledTimes(1);
-        resetDecodingInfos();
-      })
-      .catch(() => {
-        resetDecodingInfos();
-      });
+    const res = await probeDecodingInfos(configuration);
+    expect(res).toBe("Supported");
+    expect(decodingInfoStub).toHaveBeenCalledTimes(1);
+    resetDecodingInfos();
   });
 
-  it("should resolve with `Supported` if decodingInfo supports video + audio", () => {
+  it("should resolve with `Supported` if decodingInfo supports video + audio", async () => {
     const decodingInfoStub = stubDecodingInfo(true);
     const configuration: IMediaConfiguration = {
       type: "media-source",
@@ -252,19 +246,14 @@ describe("MediaCapabilitiesProber probers - decodingInfo", () => {
         channels: "5.1",
       },
     };
-    expect.assertions(2);
-    return probeDecodingInfos(configuration)
-      .then((res) => {
-        expect(res).toBe("Supported");
-        expect(decodingInfoStub).toHaveBeenCalledTimes(1);
-        resetDecodingInfos();
-      })
-      .catch(() => {
-        resetDecodingInfos();
-      });
+
+    const res = await probeDecodingInfos(configuration);
+    expect(res).toBe("Supported");
+    expect(decodingInfoStub).toHaveBeenCalledTimes(1);
+    resetDecodingInfos();
   });
 
-  it("should return `NotSupported` if no decodingInfo support (video only)", () => {
+  it("should return `NotSupported` if no decodingInfo support (video only)", async () => {
     const decodingInfoStub = stubDecodingInfo(false);
     const configuration: IMediaConfiguration = {
       type: "media-source",
@@ -276,19 +265,14 @@ describe("MediaCapabilitiesProber probers - decodingInfo", () => {
         width: 1920,
       },
     };
-    expect.assertions(2);
-    return probeDecodingInfos(configuration)
-      .then((res) => {
-        expect(res).toBe("NotSupported");
-        expect(decodingInfoStub).toHaveBeenCalledTimes(1);
-        resetDecodingInfos();
-      })
-      .catch(() => {
-        resetDecodingInfos();
-      });
+
+    const res = await probeDecodingInfos(configuration);
+    expect(res).toBe("NotSupported");
+    expect(decodingInfoStub).toHaveBeenCalledTimes(1);
+    resetDecodingInfos();
   });
 
-  it("should return `NotSupported` if no decodingInfo support (audio only)", () => {
+  it("should return `NotSupported` if no decodingInfo support (audio only)", async () => {
     const decodingInfoStub = stubDecodingInfo(false);
     const configuration: IMediaConfiguration = {
       type: "media-source",
@@ -299,19 +283,14 @@ describe("MediaCapabilitiesProber probers - decodingInfo", () => {
         channels: "5.1",
       },
     };
-    expect.assertions(2);
-    return probeDecodingInfos(configuration)
-      .then((res) => {
-        expect(res).toBe("NotSupported");
-        expect(decodingInfoStub).toHaveBeenCalledTimes(1);
-        resetDecodingInfos();
-      })
-      .catch(() => {
-        resetDecodingInfos();
-      });
+
+    const res = await probeDecodingInfos(configuration);
+    expect(res).toBe("NotSupported");
+    expect(decodingInfoStub).toHaveBeenCalledTimes(1);
+    resetDecodingInfos();
   });
 
-  it("should return `NotSupported` if no decodingInfo support", () => {
+  it("should return `NotSupported` if no decodingInfo support", async () => {
     const decodingInfoStub = stubDecodingInfo(false);
     const configuration: IMediaConfiguration = {
       type: "media-source",
@@ -329,19 +308,14 @@ describe("MediaCapabilitiesProber probers - decodingInfo", () => {
         channels: "5.1",
       },
     };
-    expect.assertions(2);
-    return probeDecodingInfos(configuration)
-      .then((res) => {
-        expect(res).toBe("NotSupported");
-        expect(decodingInfoStub).toHaveBeenCalledTimes(1);
-        resetDecodingInfos();
-      })
-      .catch(() => {
-        resetDecodingInfos();
-      });
+
+    const res = await probeDecodingInfos(configuration);
+    expect(res).toBe("NotSupported");
+    expect(decodingInfoStub).toHaveBeenCalledTimes(1);
+    resetDecodingInfos();
   });
 
-  it("should resolve with `NotSupported` if decodingInfo throws", () => {
+  it("should resolve with `NotSupported` if decodingInfo throws", async () => {
     const decodingInfoStub = stubDecodingInfo(true, true);
     const configuration: IMediaConfiguration = {
       type: "media-source",
@@ -359,15 +333,10 @@ describe("MediaCapabilitiesProber probers - decodingInfo", () => {
         channels: "5.1",
       },
     };
-    expect.assertions(2);
-    return probeDecodingInfos(configuration)
-      .then((res) => {
-        expect(res).toBe("NotSupported");
-        expect(decodingInfoStub).toHaveBeenCalledTimes(1);
-        resetDecodingInfos();
-      })
-      .catch(() => {
-        resetDecodingInfos();
-      });
+
+    const res = await probeDecodingInfos(configuration);
+    expect(res).toBe("NotSupported");
+    expect(decodingInfoStub).toHaveBeenCalledTimes(1);
+    resetDecodingInfos();
   });
 });
